Add files via upload
This is a fully functional code that can be used on the ATM. It allows a user to :- Create a bank account by supplying a user id and password, Login using their id and password & Quit the program. If login is not successful (for example the id or password did not match) then the user is taken back to the introduction menu & if login is successful the user is able to:-  Withdraw money, Deposit money, Request balance & terminating the process.